+++ lib/droplet_kit/mappings/firewall_outbound_rule_mapping.rb
@@ -1,0 +1,16 @@
+module DropletKit
+  class FirewallOutboundRuleMapping
+    include Kartograph::DSL
+
+    kartograph do
+      mapping FirewallOutboundRule
+      root_key plural: 'outbound_rules', scopes: [:read, :create, :update]
+
+      scoped :read, :create, :update do
+        property :protocol
+        property :ports
+        property :destinations
+      end
+    end
+  end
+end