+++ lib/droplet_kit/mappings/firewall_pending_change_mapping.rb
@@ -1,0 +1,16 @@
+module DropletKit
+  class FirewallPendingChangeMapping
+    include Kartograph::DSL
+
+    kartograph do
+      mapping FirewallPendingChange
+      root_key plural: 'pending_changes', scopes: [:read]
+
+      scoped :read do
+        property :droplet_id
+        property :removing
+        property :status
+      end
+    end
+  end
+end