+++ lib/droplet_kit/models/firewall_inbound_rule.rb
@@ -1,0 +1,7 @@
+module DropletKit
+  class FirewallInboundRule < BaseModel
+    attribute :protocol
+    attribute :ports
+    attribute :sources
+  end
+end