+++ lib/droplet_kit/models/firewall_outbound_rule.rb
@@ -1,0 +1,7 @@
+module DropletKit
+  class FirewallOutboundRule < BaseModel
+    attribute :protocol
+    attribute :ports
+    attribute :destinations
+  end
+end