+++ lib/droplet_kit/models/firewall_pending_change.rb
@@ -1,0 +1,7 @@
+module DropletKit
+  class FirewallPendingChange < BaseModel
+    attribute :droplet_id
+    attribute :removing
+    attribute :status
+  end
+end